[Type-o-Matic] Refactor Type-o-Matic calling process (#299)
This PR is the first step towards making a more efficient Type-o-Matic implementation. It lays the groundwork by allowing Type-o-Matic to be called with namespaces, even when generating C#. To do this, it makes several changes:
1. Add another flag to the Type-o-Matic call, `--generate=(csharp|swift)`. This allows the user to expressly decide to create Swift or C# code, instead of relying on the other arguments to decide.
2. Pull the Type-o-Matic call from the SwiftRuntimeLibrary out of msbuild and into make. **This does not change the entire SwiftRuntimeLibrary build process to make, it just pulls one part into a makefile.** This is done to enable the next change:
3. Pull all included namespaces into `common.mk`, then include namespaces in the Type-o-Matic call from SwiftRuntimeLibrary. This means that we now have access to the pertinent namespaces when generating C# code with Type-o-Matic, which is necessary for the planned refactor.

The next step after this PR is to move to a hash table holding type conversions, instead of a single function per type conversion. This approach requires knowing about namespaces from the C# side, hence these changes.

Note: some of the changes in this PR come from running Edit -> Format -> Format Document in `Program.cs`.