[CI] Ensure that we do not use relative paths. (#536)
Cheking out two or more repositories using the yaml results in a diff
behaviour regarding the value of the currenct working directory. While
when we only have one repo, the variable points to the checkout, if we
have more, it points to the dir with ALL the repos.

There are a number of scripts that are making the wrong assumption about
their location, resulting in 'strange' issues (not really strage, the
path does not exist because you are not longer where you think you are).

We fix this by:

* Allways use absolute paths via the sources dir variable.
* Add a step that will set up the root of the two repos we are
  interested to be used in the scripts.

This should make the CI less surprising in the future.